requirements.txt
python-telegram-bot==20.6
pymongo==4.5.0
python-dotenv==1.0.0
aiohttp==3.8.6
motor==3.3.1
python-magic==0.4.27
cryptography==41.0.4
fastapi==0.104.0
uvicorn==0.24.0
python-multipart==0.0.6